+++ .gitignore
@@ -5,7 +5,8 @@ __pycache__/
 test.py
 test_routes.py
 app.log
-DigiCertGlobalRootCA.crt.pem
+mysql.pem
+migrations
 # C extensions
 *.so
